% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recession_bars.R
\name{add_recession_bars}
\alias{add_recession_bars}
\title{add recession bars to plotly object}
\usage{
add_recession_bars(plot, extra = NULL)
}
\arguments{
\item{plot}{plot_ly object}

\item{extra}{extra shapes to be added to plot}
}
\description{
add recession bars to plotly object
}
\examples{
economics \%>\%
plot_ly(x = ~date,
        y = ~uempmed) \%>\%
  add_lines() \%>\%
  add_recession_bars()
}
