% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attempt_formal.R
\name{attempt_formal}
\alias{attempt_formal}
\title{tries to fix unnamed package functions}
\usage{
attempt_formal(f_path)
}
\arguments{
\item{f_path}{file path}
}
\value{
rewrites file with changes
}
\description{
tries to fix unnamed package functions
}
\examples{
attempt_formal(here::here("R","new_project.R"))

}
