% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_ge_analyzer.R
\name{auto_ge_analyzer}
\alias{auto_ge_analyzer}
\title{Auto Gene Expression Analysis}
\usage{
auto_ge_analyzer(filepath, type, colData, design, txDb, orgDb,
  keyType = "GENEID", txName = "TXNAME", lfcThreashold = 0,
  pAdjustMethod = "BH", pThreshold = 0.05, goOnt = "MF",
  goKeytype = "TAIR", keggOrg = "ath", keggKeytype = "kegg", ...)
}
\arguments{
\item{filepath}{a \code{vector} of file path of gene count}

\item{type}{method to count gene, one of HTSeq, Salmon, Sailfish,
kallisto}

\item{colData}{a \code{data.frame} containning sample information with
columns of sample name, treatment}

\item{design}{a \code{formula} which expresses how the counts for each gene
depend on the variables in \code{colData}. Many R \code{formula} are valid,
e.g, ~ group, ~ group + condition, ~genotype + treatment + genotype:
treatment.}

\item{txDb}{a \code{\link[TxDb]{GenomicFeatures}} object, which could be
downloaded from \code{\link[AnnotationHub-objects]{AnnotationHub}}.}

\item{orgDb}{orgDb, will be used for GO enrichment analysis}

\item{lfcThreashold}{a non-negative value which specifies a log2 fold change
threshold.}

\item{pAdjustMethod}{the method to use for adjusting p-values, see ?p.adjust}

\item{pThreshold}{a numeric to subset results of DESeq by pajust, default is
0.1}

\item{goOnt}{one of "MF", "BP", and "CC" subontologies. BP: biological process
CC: cellular component MF: molecular function}

\item{goKeytype}{GO keytype}

\item{keggOrg}{KEGG orgnism}

\item{keggKeytype}{KEGG keytype}

\item{...}{additional parameter}
}
\value{
Gene Expression Report
}
\description{
Auto Gene Expression Analysis
}
\author{
Xu Zhougeng
}
