% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detach_packages.R
\name{detach_packages}
\alias{detach_packages}
\title{detaches all non base packages in environment}
\usage{
detach_packages(...)
}
\value{
nothing
}
\description{
detaches all non base packages in environment
}
\examples{
detach_packages()
}
\seealso{
https://stackoverflow.com/questions/7505547/detach-all-packages-while-working-in-r
}
