% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_analyzer.R
\name{enrich_analyzer}
\alias{enrich_analyzer}
\title{All kinds of enrichment Analysis}
\usage{
enrich_analyzer(res, pThreshold, goOrg, goKeytype = "TAIR", goOnt = "MF",
  keggOrg = "ath", keggKeytype = "kegg", pAdjustMethod = "BH", ...)
}
\arguments{
\item{res}{result object from DESeq2 results}

\item{pThreshold}{p value ajust}

\item{goOrg}{GO orgnism database}

\item{goKeytype}{GO keytype}

\item{goOnt}{one of "BP", "MF", "CC" or "GO"}

\item{keggOrg}{KEGG orgnism}

\item{keggKeytype}{KEGG keytype}

\item{pAdjustMethod}{p value ajust method}
}
\description{
All kinds of enrichment Analysis
}
