% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file2dds.R
\name{file2dds}
\alias{file2dds}
\title{Building DESeqDataSet}
\usage{
file2dds(filepath, type, colData, design, txDb = NULL, keyType = "GENEID",
  txName = "TXNAME", ...)
}
\arguments{
\item{filepath}{a \code{vector} of file path of gene count}

\item{type}{method to count gene, one of HTSeq, Salmon, Sailfish,
kallisto}

\item{colData}{a \code{data.frame} containning sample information with
columns of sample name, treatment}

\item{design}{a \code{formula} which expresses how the counts for each gene
depend on the variables in \code{colData}. Many R \code{formula} are valid,
e.g, ~ group, ~ group + condition, ~genotype + treatment + genotype:
treatment.}

\item{txDb}{a \code{\link[TxDb]{GenomicFeatures}} object, which could be
downloaded from \code{\link[AnnotationHub-objects]{AnnotationHub}}.}

\item{...}{other paramter from \code{\link[tximport]{tximport}}}
}
\value{
DESeqDataSet object
}
\description{
\code{file2dds}  build \code{DESeqDataSet} base on  a vector
  input of sample file path, its counter and experiment design for use of
  \code{\link[DESeq]{DESeq2}}
}
