% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formal_args.R
\name{formal_args}
\alias{formal_args}
\title{copies and prints default function call}
\usage{
formal_args(function_name)
}
\arguments{
\item{function_name}{name of function}
}
\value{
prints function call and copies to clipboard
}
\description{
copies and prints default function call
}
\examples{
formal_args(mean)
}
