% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_commit_push.R
\name{git_commit_push}
\alias{git_commit_push}
\title{helper function for easily committing all and pushing changes to github}
\usage{
git_commit_push(message = "marginal change", ...)
}
\arguments{
\item{message}{commit message, defaults to marginal change}
}
\description{
helper function for easily committing all and pushing changes to github
}
\examples{
git_commit_push(message = "marginal change", ...)
}
