% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_load.R
\name{highlight_key_words}
\alias{highlight_key_words}
\title{registers r functions in env with highlight.js}
\usage{
highlight_key_words(...)
}
\arguments{
\item{...}{packages to load and highlight}
}
\description{
registers r functions in env with highlight.js
}
\examples{
highlight_key_words()
}
