% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_load.R
\name{highlight_load}
\alias{highlight_load}
\title{registers r functions in env with highlight.js}
\usage{
highlight_load(...)
}
\arguments{
\item{...}{packages to load and highlight}
}
\description{
registers r functions in env with highlight.js
}
\examples{
highlight_load(tidyverse,stats)
}
