% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_app.R
\name{new_app}
\alias{new_app}
\title{creates a boiler plate app in the proper folder under inst}
\usage{
new_app(app_name = NULL, remove_comments = TRUE)
}
\arguments{
\item{app_name}{name of new app}

\item{remove_comments}{logical remove app template comments}
}
\description{
creates a boiler plate app in the proper folder under inst
}
\examples{
new_app(app_name = "test", remove_comments = TRUE)
}
