% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_chaos_report.R
\name{new_chaos_report}
\alias{new_chaos_report}
\title{create an rmarkdown document based on downcute chaos theme}
\usage{
new_chaos_report(report_name, ...)
}
\arguments{
\item{report_name}{name of report}
}
\description{
create an rmarkdown document based on downcute chaos theme
}
\examples{
new_chaos_report(report_name = "test")
}
