% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{new_project}
\alias{new_project}
\title{Easy function to create R package from a function}
\usage{
new_project(
  package_name,
  author_name = "Zac Garland",
  email = "zachary.garland@outlook.com"
)
}
\arguments{
\item{package_name}{package name}

\item{author_name}{"First Last"}

\item{email}{"email to include"}
}
\description{
Easy function to create R package from a function
}
\examples{
new_project(package_name, author_name = "First Last", email = "")
}
