% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_html_to_r.R
\name{preprocess_html}
\alias{preprocess_html}
\title{pre processes html prior to formatting}
\usage{
preprocess_html(raw_html, length_of_param = 30)
}
\arguments{
\item{raw_html}{character vector of html text}

\item{length_of_param}{length of original input to use as placeholder}
}
\description{
pre processes html prior to formatting
}
