% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_html_to_r.R
\name{raw_html_to_r}
\alias{raw_html_to_r}
\title{converts raw html to an r function}
\usage{
raw_html_to_r(raw_html, length_of_param = 30)
}
\arguments{
\item{raw_html}{character vector of html text}

\item{length_of_param}{length of original input to use as placeholder}
}
\value{
html function template copied to clipboard
}
\description{
converts raw html to an r function
}
\examples{
raw_html_to_r('<div class="card" style="max-width: 18rem;"><p style="color-red;">text</p></div>', length_of_param = -1)
}
