% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_result.R
\name{save_result}
\alias{save_result}
\title{Annotate genes}
\usage{
save_result(result, path, org)
}
\arguments{
\item{result}{result from DESeq2 results}

\item{path}{which path to store data}

\item{org}{organism database, such as org.Hs.eg.db}
}
\description{
Annotate genes
}
