% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwrap_col_def.R
\name{unwrap_col_def}
\alias{unwrap_col_def}
\title{allows you to unwrap or display all of a column}
\usage{
unwrap_col_def(value, ..., px_factor = 10, style = NULL, headerStyle = NULL)
}
\arguments{
\item{value}{direct column ref using dot notation}

\item{...}{additional params to pass to reactable}

\item{px_factor}{pixel factor for col width}

\item{style}{previously defined style rules}

\item{headerStyle}{previously defined header rules}
}
\description{
allows you to unwrap or display all of a column
}
