% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_roxygen_comments.R
\name{use_roxygen_comments}
\alias{use_roxygen_comments}
\title{creates a roxygen template for functions}
\usage{
use_roxygen_comments(x, ...)
}
\arguments{
\item{x}{function name}
}
\value{
copies comment template to clipboard
}
\description{
creates a roxygen template for functions
}
\examples{
use_roxygen_comments(mean)
}
