% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_analyzer.R
\name{write_enrich}
\alias{write_enrich}
\title{Enrichment Analysis Results Output}
\usage{
write_enrich(enrichList, dirname, orgDb = NULL, column = NULL,
  keyType = NULL)
}
\arguments{
\item{enrichList}{a list return from enrich_analyzer}

\item{dirname}{directory name for storage of results}

\item{orgDb}{orgniams database for gnen annotation}

\item{column}{the columns or kinds of things that can be retrieved from the
database. As with keys, all possible columns are returned by using the
columns method.}

\item{keyType}{the keytype that matches the keys used. For the select
methods, this is used to indicate the kind of ID being used with the keys
argument. For the keys method this is used to indicate which kind of keys
are desired from keys}
}
\description{
Write the enrichment analysis results into csv files
}
