% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zg_core_libs.R
\name{zg_core_libs}
\alias{zg_core_libs}
\title{zg_core_libs
loads favorite functions}
\usage{
zg_core_libs(core = T, highlight = F)
}
\arguments{
\item{core}{condensed packages = T all favs = F}

\item{highlight}{highlights all function names within code in Rmd}
}
\description{
zg_core_libs
loads favorite functions
}
\examples{
zg_core_libs(core = T, highlight = F)
}
