% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgtools-package.R
\docType{package}
\name{zgtools-package}
\alias{zgtools-package}
\alias{zgtools}
\title{Xu Zhougeng's Toolkits}
\description{
I build this package to avoid rewriting some commonly used functions.
}
\author{
xuzhougeng
}
\keyword{internal}
