% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_ge_analyzer.R
\name{auto_ge_analyzer}
\alias{auto_ge_analyzer}
\title{Auto Gene Expression Analysis}
\usage{
auto_ge_analyzer(files, sampleInfo, design, txDb, orgDb, lfcThreashold = 0,
  pAdjustMehotd = "BH", pThreshold = 0.05, goOnt = "MF",
  goKeytype = "TAIR", keggOrg = "ath", keggKeytype = "kegg", ...)
}
\arguments{
\item{files}{file path of count matrix}

\item{sampleInfo}{sample infomation}

\item{design}{experiment design, i.e. design = ~ condition}

\item{txDb}{txDb, will be to import count matrix}

\item{orgDb}{orgDb, will be used for GO enrichment analysis}

\item{lfcThreashold}{a non-negative value which specifies a log2 fold change
threshold.}

\item{pThreshold}{a numeric to subset results of DESeq by pajust, default is
0.1}

\item{goOnt}{one of "BP", "MF", "CC" or "GO"}

\item{goKeytype}{GO keytype}

\item{keggOrg}{KEGG orgnism}

\item{keggKeytype}{KEGG keytype}

\item{...}{additional parameter}

\item{pAjustMehotd}{the method to use for adjusting p-values, see ?p.adjust}

\item{ont}{one of "MF", "BP", and "CC" subontologies. BP: biological process
CC: cellular component MF: molecular function}

\item{res}{result object from DESeq2 results}

\item{pThreshold}{p value ajust}
}
\value{
Gene Expression Report
}
\description{
Auto Gene Expression Analysis
}
\author{
Xu Zhougeng
}
