% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_plot.R
\name{eda_plot}
\alias{eda_plot}
\title{eda_plot}
\usage{
eda_plot(dds, intgroup, blind = TRUE)
}
\arguments{
\item{dds}{DESeqDataSet object}

\item{blind}{logical, whether to blind the transformation to the experimental design.}
}
\value{
plot
}
\description{
Exploratory Data Analysis Plot
}
